backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyJWT==2.10.1
groq==0.4.1
python-multipart==0.0.6
cors==1.0.1
livekit-api==1.0.5